+++ engine/cld/mcms/proposalutils/inspectors.go
